move item collection tests
